requirements.txt
django
django-rest-framework
xmltodict